+++ Source/OxyPlot/Annotations/Annotation.cs
@@ -9,14 +9,12 @@
 
 namespace OxyPlot.Annotations
 {
-    using System;
-
     using OxyPlot.Axes;
 
     /// <summary>
     /// Provides an abstract base class for annotations.
     /// </summary>
-    public abstract class Annotation : PlotElement
+    public abstract class Annotation : PlotElement, ITransposablePlotElement
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="Annotation" /> class.
@@ -24,6 +22,8 @@ namespace OxyPlot.Annotations
         protected Annotation()
         {
             this.Layer = AnnotationLayer.AboveSeries;
+            this.ClipByXAxis = true;
+            this.ClipByYAxis = true;
         }
 
         /// <summary>
@@ -48,6 +48,18 @@ namespace OxyPlot.Annotations
         /// </summary>
         /// <value>The Y axis.</value>
         public Axis YAxis { get; private set; }
+        
+        /// <summary>
+        /// Gets or sets a value indicating whether to clip the annotation by the X axis range.
+        /// </summary>
+        /// <value><c>true</c> if clipping by the X axis is enabled; otherwise, <c>false</c>.</value>
+        public bool ClipByXAxis { get; set; }
+
+        /// <summary>
+        /// Gets or sets a value indicating whether to clip the annotation by the Y axis range.
+        /// </summary>
+        /// <value><c>true</c> if clipping by the Y axis is enabled; otherwise, <c>false</c>.</value>
+        public bool ClipByYAxis { get; set; }
 
         /// <summary>
         /// Gets or sets the Y axis key.
@@ -60,13 +72,8 @@ namespace OxyPlot.Annotations
         /// </summary>
         public void EnsureAxes()
         {
-            this.XAxis = this.XAxisKey != null ?
-                         this.PlotModel.GetAxis(this.XAxisKey) :
-                         this.PlotModel.DefaultXAxis;
-
-            this.YAxis = this.YAxisKey != null ?
-                         this.PlotModel.GetAxis(this.YAxisKey) :
-                         this.PlotModel.DefaultYAxis;
+            this.XAxis = this.XAxisKey != null ? this.PlotModel.GetAxis(this.XAxisKey) : this.PlotModel.DefaultXAxis;
+            this.YAxis = this.YAxisKey != null ? this.PlotModel.GetAxis(this.YAxisKey) : this.PlotModel.DefaultYAxis;
         }
 
         /// <summary>
@@ -78,60 +85,38 @@ namespace OxyPlot.Annotations
         }
 
         /// <summary>
-        /// Transforms the specified coordinates to a screen point.
-        /// </summary>
-        /// <param name="x">The x coordinate.</param>
-        /// <param name="y">The y coordinate.</param>
-        /// <returns>A screen point.</returns>
-        public ScreenPoint Transform(double x, double y)
-        {
-            return this.XAxis.Transform(x, y, this.YAxis);
-        }
-
-        /// <summary>
-        /// Transforms the specified data point to a screen point.
-        /// </summary>
-        /// <param name="p">The point.</param>
-        /// <returns>A screen point.</returns>
-        public ScreenPoint Transform(DataPoint p)
-        {
-            return this.XAxis.Transform(p.X, p.Y, this.YAxis);
-        }
-
-        /// <summary>
-        /// Transforms the specified screen position to a data point.
-        /// </summary>
-        /// <param name="position">The position.</param>
-        /// <returns>A data point</returns>
-        public DataPoint InverseTransform(ScreenPoint position)
-        {
-            return this.XAxis.InverseTransform(position.X, position.Y, this.YAxis);
-        }
-
-        /// <summary>
-        /// Tests if the plot element is hit by the specified point.
-        /// </summary>
-        /// <param name="args">The hit test arguments.</param>
-        /// <returns>
-        /// A hit test result.
-        /// </returns>
-        protected override HitTestResult HitTestOverride(HitTestArguments args)
-        {
-            return null;
-        }
-
-        /// <summary>
         /// Gets the clipping rectangle.
         /// </summary>
-        /// <returns>The clipping rectangle.</returns>
+        /// <returns>
+        /// The clipping rectangle.
+        /// </returns>
         protected OxyRect GetClippingRect()
         {
-            double minX = Math.Min(this.XAxis.ScreenMin.X, this.XAxis.ScreenMax.X);
-            double minY = Math.Min(this.YAxis.ScreenMin.Y, this.YAxis.ScreenMax.Y);
-            double maxX = Math.Max(this.XAxis.ScreenMin.X, this.XAxis.ScreenMax.X);
-            double maxY = Math.Max(this.YAxis.ScreenMin.Y, this.YAxis.ScreenMax.Y);
+            var rect = this.PlotModel.PlotArea;
+            var axisRect = TransposablePlotElementExtensions.GetClippingRect(this);
 
-            return new OxyRect(minX, minY, maxX - minX, maxY - minY);
+            var minX = double.NegativeInfinity;
+            var maxX = double.PositiveInfinity;
+            var minY = double.NegativeInfinity;
+            var maxY = double.PositiveInfinity;
+
+            if (this.ClipByXAxis)
+            {
+                minX = this.Orientate(axisRect.TopLeft).X;
+                maxX = this.Orientate(axisRect.BottomRight).X;
+            }
+
+            if (this.ClipByYAxis)
+            {
+                minY = this.Orientate(axisRect.TopLeft).Y;
+                maxY = this.Orientate(axisRect.BottomRight).Y;
+            }
+
+            var minPoint = this.Orientate(new ScreenPoint(minX, minY));
+            var maxPoint = this.Orientate(new ScreenPoint(maxX, maxY));
+
+            var axisClipRect = new OxyRect(minPoint, maxPoint);
+            return rect.Clip(axisClipRect);
         }
     }
 }