+++ Source/OxyPlot/Annotations/PathAnnotation.cs
@@ -11,6 +11,7 @@ namespace OxyPlot.Annotations
 {
     using System;
     using System.Collections.Generic;
+    using OxyPlot;
 
     /// <summary>
     /// Provides an abstract base class for all annotations that contain paths (lines, functions or polylines).
@@ -35,8 +36,6 @@ namespace OxyPlot.Annotations
             this.StrokeThickness = 1;
             this.LineStyle = LineStyle.Dash;
             this.LineJoin = LineJoin.Miter;
-            this.ClipByXAxis = true;
-            this.ClipByYAxis = true;
             this.Aliased = false;
 
             this.TextLinePosition = 1;
@@ -124,18 +123,6 @@ namespace OxyPlot.Annotations
         public bool ClipText { get; set; }
 
         /// <summary>
-        /// Gets or sets a value indicating whether to clip the annotation line by the X axis range.
-        /// </summary>
-        /// <value><c>true</c> if clipping by the X axis is enabled; otherwise, <c>false</c>.</value>
-        public bool ClipByXAxis { get; set; }
-
-        /// <summary>
-        /// Gets or sets a value indicating whether to clip the annotation line by the Y axis range.
-        /// </summary>
-        /// <value><c>true</c> if clipping by the Y axis is enabled; otherwise, <c>false</c>.</value>
-        public bool ClipByYAxis { get; set; }
-
-        /// <summary>
         /// Gets or sets a value indicating whether the path is aliased.
         /// </summary>
         /// <value><c>true</c> if is aliased; otherwise, <c>false</c>.</value>
@@ -177,12 +164,7 @@ namespace OxyPlot.Annotations
 
             this.screenPoints = this.GetScreenPoints();
 
-            // clip to the area defined by the axes
-            var clippingRectangle = OxyRect.Create(
-                this.ClipByXAxis ? this.XAxis.ScreenMin.X : this.PlotModel.PlotArea.Left,
-                this.ClipByYAxis ? this.YAxis.ScreenMin.Y : this.PlotModel.PlotArea.Top,
-                this.ClipByXAxis ? this.XAxis.ScreenMax.X : this.PlotModel.PlotArea.Right,
-                this.ClipByYAxis ? this.YAxis.ScreenMax.Y : this.PlotModel.PlotArea.Bottom);
+            var clippingRectangle = this.GetClippingRect();
 
             const double MinimumSegmentLength = 4;
 
@@ -204,11 +186,11 @@ namespace OxyPlot.Annotations
                    clippedPoints.AddRange);
             }
 
-            ScreenPoint position;
-            double angle;
-            double margin = this.TextMargin;
+            var margin = this.TextMargin;
 
-            if (this.TextHorizontalAlignment == HorizontalAlignment.Center)
+            this.GetActualTextAlignment(out var ha, out var va);
+
+            if (ha == HorizontalAlignment.Center)
             {
                 margin = 0;
             }
@@ -217,7 +199,7 @@ namespace OxyPlot.Annotations
                 margin *= this.TextLinePosition < 0.5 ? 1 : -1;
             }
 
-            if (GetPointAtRelativeDistance(clippedPoints, this.TextLinePosition, margin, out position, out angle))
+            if (GetPointAtRelativeDistance(clippedPoints, this.TextLinePosition, margin, out var position, out var angle))
             {
                 if (angle < -90)
                 {
@@ -243,12 +225,12 @@ namespace OxyPlot.Annotations
                 var angleInRadians = angle / 180 * Math.PI;
                 var f = 1;
 
-                if (this.TextHorizontalAlignment == HorizontalAlignment.Right)
+                if (ha == HorizontalAlignment.Right)
                 {
                     f = -1;
                 }
 
-                if (this.TextHorizontalAlignment == HorizontalAlignment.Center)
+                if (ha == HorizontalAlignment.Center)
                 {
                     f = 0;
                 }
@@ -292,8 +274,8 @@ namespace OxyPlot.Annotations
                            this.ActualFontSize,
                            this.ActualFontWeight,
                            angle,
-                           this.TextHorizontalAlignment,
-                           this.TextVerticalAlignment);
+                           ha,
+                           va);
                     }
                 }
             }
@@ -339,20 +321,19 @@ namespace OxyPlot.Annotations
             this.ActualMinimumY = Math.Max(this.MinimumY, this.YAxis.ActualMinimum);
             this.ActualMaximumY = Math.Min(this.MaximumY, this.YAxis.ActualMaximum);
 
+            var topLeft = this.InverseTransform(this.PlotModel.PlotArea.TopLeft);
+            var bottomRight = this.InverseTransform(this.PlotModel.PlotArea.BottomRight);
+
             if (!this.ClipByXAxis)
             {
-                double right = this.XAxis.InverseTransform(this.PlotModel.PlotArea.Right);
-                double left = this.XAxis.InverseTransform(this.PlotModel.PlotArea.Left);
-                this.ActualMaximumX = Math.Max(left, right);
-                this.ActualMinimumX = Math.Min(left, right);
+                this.ActualMaximumX = Math.Max(topLeft.X, bottomRight.X);
+                this.ActualMinimumX = Math.Min(topLeft.X, bottomRight.X);
             }
 
             if (!this.ClipByYAxis)
             {
-                double bottom = this.YAxis.InverseTransform(this.PlotModel.PlotArea.Bottom);
-                double top = this.YAxis.InverseTransform(this.PlotModel.PlotArea.Top);
-                this.ActualMaximumY = Math.Max(top, bottom);
-                this.ActualMinimumY = Math.Min(top, bottom);
+                this.ActualMaximumY = Math.Max(topLeft.Y, bottomRight.Y);
+                this.ActualMinimumY = Math.Min(topLeft.Y, bottomRight.Y);
             }
         }